pyproject.toml
[build-system]
# Minimum requirements for the build system to execute.
requires = ["setuptools", "wheel"]  # PEP 508 specifications.

[project]
version = "1.0.0"
name = "VSPEC"
authors = [
    { name = "Ted Johnson", email = "ted.johnson@unlv.edu" },
    { name = "Cameron Kelahan", email = "cameron.kelahan@gmail.com"}
]
maintainers = [ { name = "Ted Johnson", email = "ted.johnson@unlv.edu" } ]
description = "Simulate Observations of Exoplanets in Variable Star Systems"
readme = "docs/readme.rst"
license = {file = "LICENSE"}
keywords = ["exoplanet", "astronomy", "star"]
requires-python = ">= 3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy~=1.26.3",
    "astropy",
    "scipy",
    "tqdm",
    "matplotlib",
    "pathlib",
    "netCDF4",
    "pyyaml",
    "libpypsg~=0.4.0",
    "vspec-vsm~=0.3.7",
    "GridPolator~=0.4.1",
]
[project.optional-dependencies]
plot = ["matplotlib","cartopy"]
dev = [
"pytest",
"pytest-timeout",
"pandas",
"pep8",
"pydata-sphinx-theme",
"imageio",
"numpydoc",
"sphinx",
"sphinx-automodapi",
"sphinx-gallery",
"sphinxcontrib-bibtex",
"sphinxcontrib-napoleon",
]

[project.urls]
Documentation = "https://vspec-collab.github.io/VSPEC/1.0.0/"
Repository = "https://github.com/VSPEC-collab/VSPEC"
Issues = "https://github.com/VSPEC-collab/VSPEC/issues"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "plot: marks tests as requiring plotting (deselect with '-m \"not plot\"')",
]

[tool.setuptools.package-data]
"*" = ["*.json","*.yaml"]